Fix 1x/evdo data call for NV based subscription.
(PeterCxy: this is not a duplicate. Last commit with the same title was lost when I cherry-picked some code today, so do this again)
NV based CDMA subscription was not handled properly in
DcTracker and data calls were not initiated since it was
waiting for sim loaded event.Changed the code to allow data
calls even without loading SIM ,if CDMA subscription is NV
based.
Also for 1x/evdo data calls with non OMH cards, changed the
code to use dummy APN profiles instead of the APNs loaded
based on SIM.
Dummy APN profiles will be created based on the current DATA
radio technology so that they wont be used for GSM/UMTS/LTE
and eHRPD.

CRs-fixed: 518900, 517763
Change-Id: I815d2110d469b646c7f5905f5bfc6fd14e6251b7

Conflicts:
	src/java/com/android/internal/telephony/Phone.java
	src/java/com/android/internal/telephony/dataconnection/DcTracker.java